Starter module loading, no eval, move to helper.
