+++ tests/starter/test_starter_helper.py
@@ -3,6 +3,7 @@ import starter.starter_helper as starter_helper
 import tests.test_data as test_data
 import json
 from ddt import ddt, data, unpack
+from tests.activity.classes_mock import FakeLogger
 
 
 example_workflow_name = "PostPerfectPublication"
@@ -38,7 +39,23 @@ class TestStarterHelper(unittest.TestCase):
         self.assertEqual("1800", execution_start_to_close_timeout)
         self.assertEqual(json.dumps(data), workflow_input)
 
+    def test_get_starter_module(self):
+        "for coverage successful starter module import"
+        starter_name = 'starter_PackagePOA'
+        module_object = starter_helper.get_starter_module(starter_name, FakeLogger())
+        self.assertIsNotNone(module_object)
 
+    def test_get_starter_module_failure(self):
+        "for coverage test failure"
+        starter_name = 'not_a_starter'
+        module_object = starter_helper.get_starter_module(starter_name, FakeLogger())
+        self.assertIsNone(module_object)
+
+    def test_import_starter_module_failure(self):
+        "for coverage test import failure"
+        starter_name = 'not_a_starter'
+        return_value = starter_helper.import_starter_module(starter_name, FakeLogger())
+        self.assertFalse(return_value)
 
 if __name__ == '__main__':
     unittest.main()
